Finished eula screen, improved frames/states
Merged frames and states together, simplifying the logic for frames and states. Overall, easier to maintain and add new screens.

Eula screen finished.
