--- src/nStaller/Frames/AgreementFrame.cpp
@@ -1,75 +1,0 @@
-#include "AgreementFrame.h"
-#include "../Installer.h"
-
-
-static LRESULT CALLBACK WndProc(HWND, UINT, WPARAM, LPARAM);
-
-AgreementFrame::~AgreementFrame()
-{
-	UnregisterClass("AGREEMENT_FRAME", m_hinstance);
-	DestroyWindow(m_hwnd);
-	DestroyWindow(m_checkNo);
-	DestroyWindow(m_checkYes);
-}
-
-AgreementFrame::AgreementFrame(Installer * installer, const HINSTANCE hInstance, const HWND parent, const RECT & rc)
-{
-	// Create window class
-	m_installer = installer;
-	m_hinstance = hInstance;
-	m_wcex.cbSize = sizeof(WNDCLASSEX);
-	m_wcex.style = CS_HREDRAW | CS_VREDRAW;
-	m_wcex.lpfnWndProc = WndProc;
-	m_wcex.cbClsExtra = 0;
-	m_wcex.cbWndExtra = 0;
-	m_wcex.hInstance = hInstance;
-	m_wcex.hIcon = LoadIcon(hInstance, IDI_APPLICATION);
-	m_wcex.hCursor = LoadCursor(NULL, IDC_ARROW);
-	m_wcex.hbrBackground = (HBRUSH)(COLOR_WINDOW + 1);
-	m_wcex.lpszMenuName = NULL;
-	m_wcex.lpszClassName = "AGREEMENT_FRAME";
-	m_wcex.hIconSm = LoadIcon(m_wcex.hInstance, IDI_APPLICATION);
-	RegisterClassEx(&m_wcex);
-	m_hwnd = CreateWindow("AGREEMENT_FRAME", "", WS_OVERLAPPED | WS_VISIBLE | WS_CHILD, rc.left, rc.top, rc.right - rc.left, rc.bottom - rc.top, parent, NULL, hInstance, NULL);
-	SetWindowLongPtr(m_hwnd, GWLP_USERDATA, (LONG_PTR)this);
-	setVisible(false);
-
-
-}
-
-static LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
-{
-	auto ptr = (AgreementFrame*)GetWindowLongPtr(hWnd, GWLP_USERDATA);
-	if (message == WM_PAINT) {
-		PAINTSTRUCT ps;
-		Graphics graphics(BeginPaint(hWnd, &ps));
-
-		// Draw Background
-		LinearGradientBrush backgroundGradient(
-			Point(0, 0),
-			Point(0, 450),
-			Color(50, 25, 125, 225),
-			Color(255, 255, 255, 255)
-		);
-		graphics.FillRectangle(&backgroundGradient, 0, 0, 630, 450);
-
-		// Preparing Fonts
-		FontFamily  fontFamily(L"Segoe UI");
-		Font        bigFont(&fontFamily, 25, FontStyleBold, UnitPixel);
-		Font        regFont(&fontFamily, 14, FontStyleRegular, UnitPixel);
-		SolidBrush  blueBrush(Color(255, 25, 125, 225));
-		SolidBrush  blackBrush(Color(255, 0, 0, 0));
-
-		// Draw Text
-		graphics.SetSmoothingMode(SmoothingMode::SmoothingModeAntiAlias);
-		graphics.DrawString(L"License Agreement", -1, &bigFont, PointF{ 10, 10 }, &blueBrush);
-		graphics.DrawString(L"Please read the following license agreement.", -1, &regFont, PointF{ 10, 100 }, &blackBrush);
-	
-
-		EndPaint(hWnd, &ps);
-		return S_OK;
-	}
-	else if (message == WM_COMMAND) {
-	}	
-	return DefWindowProc(hWnd, message, wParam, lParam);
-}
