+++ src/nStaller/manifest.nman
@@ -1,0 +1,11 @@
+name "Empty"
+
+version "1.2.3.4"
+
+description
+"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum." 
+
+eula
+"nSuite installers can be created freely by anyone, as such those who generate them are responsible for its contents, not the developers.
+
+This software is provided as-is, use it at your own risk."
